QA_Chatbot/requirements.txt
streamlit
google-generativeai
python-dotenv
os